Unroll inner loop. Tasks instead of threads.
The naive thread method from the reference C code completeley
overwhelmed the timing. The naive Task-based method here are a lot
better but still rediculously slow. The solution is probably to setup
the work items for the tasks, create the requested number of tasks - 1
once and have the tasks + the main thread pull work items and process
them until all the data is processed.